generate cmake
