+++ templates/cmake.tmpl
@@ -1,0 +1,37 @@
+// cmake generated by go generate; DO NOT EDIT.
+// This file was generated by robots at
+// {{ .Timestamp }}
+
+cmake_minimum_required(VERSION 3.1)
+
+
+include_directories(
+                   ${PROJECT_SOURCE_DIR}"
+                   )
+
+add_executable(
+    {{ .Executable }} 
+    {{ .Executable }}.cpp
+)
+
+if (WIN32) 
+    target_compile_definitions({{ .Executable }} PRIVATE JSON_SO=1 _LIB JSON_COMPILING=1)
+    target_compile_definitions(jsonv PRIVATE JSON_SO=1 _LIB JSON_COMPILING=1)
+endif(WIN32)
+
+target_link_libraries(
+    {{ .Executable }} 
+    crud_service
+    vanilla_rtb
+    jsonv
+    parsers
+    ${Boost_LIBRARIES}
+    ${CMAKE_THREAD_LIBS_INIT}
+    ${RT_LIB}
+)
+
+install(TARGETS {{ .Executable }} 
+  RUNTIME DESTINATION bin
+  LIBRARY DESTINATION lib
+)
+
